% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{add_customer_entry}
\alias{add_customer_entry}
\alias{search_customer}
\alias{view_customer}
\title{Add customers to a library}
\usage{
add_customer_entry(company, name, street, city, country, VAT)

search_customer(keywords, search = "|")

view_customer(...)
}
\arguments{
\item{company}{Character string for company name.}

\item{name}{Character string for contact name.}

\item{street}{Character string for street name.}

\item{city}{Character string for city name.}

\item{country}{Character string for country name.}

\item{VAT}{Character string for VAT number.}

\item{keywords}{Keywords for searching the database.}

\item{search}{Determine whether multiple keywords are searched with the
boolean \code{"|"}, or \code{"&"} for an exclusive combination of keywords (default =
\code{"|"}).}

\item{...}{Currently not supported.}
}
\value{
\code{tibble::\link[tibble:tibble]{tibble}()}
}
\description{
\code{add_customer_entry()} creates a library of your customer while assigning a
unique ID, or customer number. This is used when creating invoices with
\code{render_invoice()}. Search entries with \code{search_customer()} and view the
whole database with \code{view_customer()}.
}
\examples{
\dontrun{
# customer 1
add_customer_entry(
 "Fancy Stuff inc",
 "G. Janssen",
 "Liftweg 135",
 "7009 ZZ Doetinchem",
 "Netherlands",
 "NL10937277"
)
# customer 2
add_customer_entry(
 "Fancy Work",
 NA_character_,
 "Vleutensebaan 435",
 "8888 ZZ Haarlem",
 "Netherlands",
 "NL1093se44"
 }
}
