% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybook.R
\name{add_daybook_entry}
\alias{add_daybook_entry}
\title{Add daybook entries}
\usage{
add_daybook_entry(
  date,
  daybook,
  account,
  entity,
  comment,
  invoice_file,
  invoice_num,
  direction,
  currency,
  amount,
  ...,
  VAT_class
)
}
\arguments{
\item{date}{Date of transaction.}

\item{daybook}{A character string for the daybook type (sales, purchases,
bank, memorial).}

\item{account}{A numeric value for ledger account number.}

\item{entity}{Describe the transaction (i.e., the company).}

\item{comment}{Describe in more detail the transaction.}

\item{invoice_file}{The path to the invoice file.}

\item{invoice_num}{The invoice number as given on the invoice as a numeric
value.}

\item{direction}{The direction of the transaction, usually \code{"credit"}, when
using the function as a stand-alone.}

\item{currency}{A character string for the currency of the transaction.}

\item{amount}{A numeric value for the amount of the transaction.}

\item{...}{A numeric value or vector of numeric values for the absolute
value(s) of VAT included in the invoice.}

\item{VAT_class}{A variable for the VAT class of the transaction. Supply a
character vector with the same number of element as VAT object supplied to
the ellipses (\code{...}).}
}
\description{
This function is mainly used as a standalone when adding invoices from a
creditor to you administration.
}
\examples{
\dontrun{
add_daybook_entry(
 Sys.Date(),
 "sales",
 1050,
 "Super-store",
 "A bunch of things",
 "invoice.txt",
 1,
 "credit",
 "$",
 1000,
 210,
 "high-VAT"
)
}
}
