% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{add_entry}
\alias{add_entry}
\title{Add journal entry}
\usage{
add_entry(
  date,
  currency,
  amount,
  VAT,
  direction,
  account_ref,
  entity,
  comment,
  invoice_file,
  invoice_num
)
}
\description{
Add journal entry
}
