% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book-profile.R
\name{create_accountant}
\alias{create_accountant}
\title{Create an accountant account}
\usage{
create_accountant(.text = NULL, quiet = FALSE)
}
\arguments{
\item{.text}{Vector of character string of length 11. In consecutive order
listing; company name, street and street number, zip code and city, country,
email address, phone number, website, VAT ID, VAT tax number, IBAN and BIC.}

\item{quiet}{Prevents messages from being printed to the console (default =
\code{FALSE}).}
}
\value{
\code{yml} file
}
\description{
This creates an user account for a \code{bookkeeper} project with details
concerning the company, such as; address, banking and tax details. By default
the \code{yml} file is added to an \code{gitignore} file, if available, to
prevent leakage of sensitive information.
}
\examples{
\dontrun{
# By reading data from the console
create_accountant()

# By adding a vector of length 11 with company name, street and street number
, zip code and city, country, email address, phone number, website, VAT ID,
VAT tax number, IBAN and BIC.
text <- c(
 "FAIReLABS",
 "Modelstraat 12",
 "3017 KH Amsterdam",
 "Netherlands",
 "schobbenmartin\@gmail.com",
 "03012345",
 "https://martinschobben.github.io/webpage/",
 "12345678",
 "NL123456789B01",
 "NL99ABCD0123456789",
 "AAAABBCCDD"
)

create_accountant(.text = text, quiet = TRUE)
}
}
