% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-accounting.R
\name{render_accounting}
\alias{render_accounting}
\title{Render the account reports}
\usage{
render_accounting(template)
}
\arguments{
\item{template}{Type of output}
}
\value{

}
\description{
Render the account reports
}
