% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-invoice.R
\name{render_invoice}
\alias{render_invoice}
\alias{add_bill_entry}
\alias{make_bill}
\alias{invoice_document}
\title{Create an invoice}
\usage{
render_invoice(
  customer_num,
  bill = NULL,
  lang = "en",
  account = search_account_ref("Handelsdebiteuren")$`reference number`,
  open_doc = rlang::is_interactive(),
  quiet = FALSE
)

add_bill_entry(description, VAT, currency, price, .group = "items")

make_bill(currency_out = intToUtf8(8364), lang = "en")

invoice_document(..., keep_tex = FALSE)
}
\arguments{
\item{customer_num}{Customer number as a six digit character string.}

\item{bill}{Default is set to \code{NULL} but the compiled bill can be provided as
a \code{data.frame()} or \code{tibble::\link[tibble:tibble]{tibble}()}. However,
\code{add_bill_entry} and \code{make_bill()} should be used to generate the \code{tibble}.}

\item{lang}{Language, defaults to English (\code{"en"}).}

\item{account}{The ledger account number as a character string. The
respective account number for country specific standard codes can be
looked-up with \code{search_account_ref()}. Currently defaults to the Dutch
\href{https://www.referentiegrootboekschema.nl/ondernemers/werken-met-rgs}{RGS}
code for debtors.}

\item{open_doc}{Should the Rmd file be opened
(default = \code{rlang::\link[rlang:is_interactive]{is_interactive}()}).}

\item{quiet}{Should message not be printed (default = \code{FALSE}).}

\item{description}{Add a description of the item.}

\item{VAT}{Add VAT as a percentage to bill entry.}

\item{currency}{Add currency as to bill entry.}

\item{price}{Add price to bill entry.}

\item{.group}{Add grouping key to bill entry (default = \code{"items"}).}

\item{currency_out}{Currency to be used on the bill (default = \code{"€"}).}

\item{...}{Additional arguments supplied to Rmarkdown.}

\item{keep_tex}{Keep the tex file.}
}
\value{
A \code{komaletter::\link[komaletter:komaletter]{komaletter}} or
\code{tibble::\link[tibble:tibble]{tibble}}.
}
\description{
Generate an \code{komaletter::\link[komaletter:komaletter]{komaletter}} type
invoice by first compiling all cost posts in a bill with \code{add_bill_entry()},
and finishing the bill with \code{make_bill()}, before rendering the invoice with
\code{render_invoice()}. You can search for the customer number with the
\code{search_customer()} in your own customer database.
}
\examples{

\dontrun{
# make customer database
add_customer_entry(
 "Fancy Work",
 NA_character_,
 "Vleutensebaan 435",
 "7009 ZZ Haarlem",
 "Netherlands",
 "NL1093se44"
 )

# initiate bill
add_bill_entry(
 "Tekstproductie (80 uur à € 70)",
 21,
 "€",
 5600
)

# entry
add_bill_entry(
  "Ontwerp (vaste prijs)",
  21,
  "€",
  1250
  )

# entry
add_bill_entry(
  "10 foto’s à € 150",
  21,
  "€",
  1500
 )

# entry
add_bill_entry(
 "Vormgeving (30 uur à € 60)",
 21,
 "€",
 1800
)

# entry
add_bill_entry(
 "1 kg Suiker",
 9,
 "€",
 3000
)

# compile bille
make_bill(lang = "nl")

# Look-up customers (if none exist, first create a customer database)
view_customer()

# render invoice
render_invoice("060101", lang = "nl")
}

}
