% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_path.R
\name{rmd_template_path}
\alias{rmd_template_path}
\title{Get path to Rmd template}
\usage{
rmd_template_path(template = "invoice-en", type = "skeleton", path = NULL)
}
\arguments{
\item{template}{The name of the template.}

\item{type}{Type of file Rmd \code{skeleton} or \code{resources} tex files.}

\item{path}{Name of file. If \code{NULL}, the Rmarkdown skeleton file is listed.}
}
\description{
This function comes from the package \code{readr}, and has been modified to access
the bundled datasets in directory \code{inst/rmarkdown/templates} of \code{bookkeeper}.
This function make them easy to access. This function is modified from
\code{\link[readr:readr_example]{readr_example}} of the package
\code{\link[readr]{readr}}.
}
\examples{
rmd_template_path()
rmd_template_path("2018-01-19-GLENDON")
}
