% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{add_entry}
\alias{add_entry}
\title{Add journal entry}
\usage{
add_entry(
  date,
  currency,
  amount,
  VAT,
  direction,
  account_ref,
  entity,
  comment,
  invoice_file,
  invoice_num
)
}
\description{
Add journal entry
}
\examples{
add_entry(Sys.Date(), "$", 1000, 200, "credit", "BFvaLenLenLn1", "The bank",
          "start-up", "invoice.pdf", 1)
}
