% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-codes.R
\name{check_account_ref}
\alias{check_account_ref}
\alias{search_account_ref}
\title{Family of functions to query reference codes}
\usage{
check_account_ref(account_ref)

search_account_ref(keywords, search = "|", type = c("C", "D"))
}
\arguments{
\item{account_ref}{A character string of the account reference code.}

\item{keywords}{A character string or vector to query the data base reference
code description fields.}

\item{search}{A character indicating whether the search is inclusive ("|") or
exclusive ("&").}

\item{type}{A character string or vector indicating whether debit ("D") or
credit ("C) is consulted.}
}
\value{
One or more entries form the coder reference database in tibble
format.
}
\description{
Family of functions to query reference codes
}
\examples{
# Use code reference to find description
check_account_ref("BFvaOvrVgbBej")
# Use keywords to find a code reference that reflect the invoice type
search_account_ref(c("lening", "mutaties"), "&", type =c("C", "D")
}
