% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-accounting.R
\name{render_accounting}
\alias{render_accounting}
\title{Render the account reports}
\usage{
render_accounting(type, template, output = "bookdown::pdf_book")
}
\arguments{
\item{type}{Type of journal}

\item{template}{Type of output}

\item{output}{Type of format}
}
\value{
Rmarkdown file
}
\description{
Render the account reports
}
